Began Comm Code
Added initilization of communication code. The rest will be finished in the next couple of days. @jardthebard The place where serverIP is one of the places the GUI should interface for setting the esu bridge IP adress. The port should just be the default unless the train club modified that but we can worry about that later.